gjhieog

+++ client/app/data.py
@@ -8,9 +8,12 @@ conf_lock = threading.Lock()
 
 port = int(sys.argv[1]) if (len(sys.argv) > 1 and sys.argv[1][0]!='-') else 9999
 data_dir = "data_"+str(port)
-if '--clear' in sys.argv:
+if '--clear' in sys.argv and os.path.exists(data_dir):
     shutil.rmtree(data_dir)
+
 default_dir = os.path.expanduser("~/Desktop/Backdown")
+if '--clear' in sys.argv and os.path.exists(default_dir):
+    shutil.rmtree(default_dir)
 if not os.path.exists(data_dir):
     os.mkdir(data_dir)
 conf_path = os.path.join(data_dir, 'conf.json')
@@ -27,3 +30,5 @@ def save_conf(data):
     conf_lock.acquire()
     open(conf_path, 'w').write(json.dumps(data))
     conf_lock.release()
+
+
